Move jira fields filtering logic in a separate file
